+++ route/search/search.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+module.exports = router;
+
+router.get("/", (req, res) => {
+  res.json("hello world");
+});
